Reorganize modules so that `from saw import *` brings in "easy" version
