+++ saw-remote-api/test-scripts/salsa20_easy.py
@@ -1,9 +1,8 @@
 import os
 import os.path
 from cryptol.cryptoltypes import to_cryptol
-import saw
+from saw import *
 from saw.llvm import Contract, LLVMArrayType, uint8_t, uint32_t, void
-from saw.easy import *
 
 dir_path = os.path.dirname(os.path.realpath(__file__))
 connect("cabal new-exec --verbose=0 saw-remote-api -- --dynamic4")
